[config-plugins] move XML.ts to utils
