+++ packages/config-plugins/src/android/Manifest.ts
@@ -1,8 +1,8 @@
 import fs from 'fs-extra';
 import path from 'path';
 
+import * as XML from '../utils/XML';
 import { assert } from '../utils/errors';
-import * as XML from './XML';
 
 export type StringBoolean = 'true' | 'false';
