+++ packages/config-plugins/src/plugins/compiler-plugins.ts
@@ -15,11 +15,11 @@ import { AndroidManifest } from '../android/Manifest';
 import * as AndroidPaths from '../android/Paths';
 import { readResourcesXMLAsync, ResourceXML } from '../android/Resources';
 import { getProjectStringsXMLPathAsync } from '../android/Strings';
-import { writeXMLAsync } from '../android/XML';
 import { getEntitlementsPath } from '../ios/Entitlements';
 import { InfoPlist } from '../ios/IosConfig.types';
 import { AppDelegateProjectFile, getAppDelegate, getInfoPlistPath } from '../ios/Paths';
 import { getPbxproj } from '../ios/utils/Xcodeproj';
+import { writeXMLAsync } from '../utils/XML';
 import { assert } from '../utils/errors';
 import * as WarningAggregator from '../utils/warnings';
 import { withInterceptedMod } from './core-plugins';
