+++ packages/config-plugins/src/utils/XML.ts
